refactor: 검증용 클래스 static으로 사용
