--- java-playground/src/main/java/org/gonza/javaplayground/Reader.java
@@ -1,7 +1,0 @@
-package org.gonza.javaplayground;
-
-interface Reader {
-    String read();
-
-    String readWithoutValidation();
-}
