+++ java-playground/src/main/java/org/gonza/javaplayground/view/Reader.java
@@ -1,0 +1,7 @@
+package org.gonza.javaplayground.view;
+
+public interface Reader {
+    String read();
+
+    String readWithoutValidation();
+}
